requirements.txt
# Required for GPU support on Windows
--extra-index-url https://download.pytorch.org/whl/cu121

# Core ML Libraries (Locked to GPU-compatible versions)
torch==2.2.0+cu121
torchtext==0.17.0
torchdata==0.7.1
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.15.3

# Utilities
PyYAML==6.0.3
tqdm==4.67.1
tensorboard==2.20.0
pillow==12.0.0
requests==2.32.5
nvidia-ml-py==13.590.44
portalocker==3.2.0

# Visualization & Reporting
matplotlib==3.8.2
pandas==2.1.4